refactor: split tunnel socks relay helpers
